pyproject.toml
[project]
name = "project-olx"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "channels>=4.3.2",
    "channels-redis>=4.3.0",
    "daphne>=4.2.1",
    "django>=6.0.3",
    "django-filter>=25.2",
    "django-json-widget>=2.1.1",
    "django-mptt>=0.18.0",
    "fixtures>=4.3.2",
    "flake8>=7.3.0",
    "pillow>=12.2.0",
    "psycopg2-binary>=2.9.11",
    "requests>=2.33.1",
]

[dependency-groups]
dev = [
    "dotenv>=0.9.9",
]
